Attempt to switch to YAML form of pipeline invocation.
I'm not sure this is correct - ci_hsc_gen3's override pipeline seems
to have some configs that ought to be important, but also shouldn't be
in ci_hsc_gen3 under that logic.  But what's present now is clearly
wrong, and testing its replacement is out of scope for this ticket.