Write the deepCoadd_skyMap dataset at the end of bootstrap.
We can't rely on the gen2to3 conversion code to add this, because it's
not actually present in the root repo we're converted in Gen2 (we just
assume in the brightObjectMask datasets that _are_ there that this
particular default HSC skymap is the one they're defined on.

But that code has to actually register the skymap in terms of
dimensions, and that means we can't just run MakeGen3SkyMapTask to put
the skymap dataset.  That should be fixed on DM-24355 (by allowing the
same skymap to be registered multiple times).